codeclone_detection: Move to c_languages
It's using clang and thus only viable for C stuff.
