+++ bears/c_languages/codeclone_detection/ClangCloneDetectionBear.py
@@ -1,8 +1,8 @@
+from bears.c_languages.codeclone_detection.ClangFunctionDifferenceBear import (
+    ClangFunctionDifferenceBear)
+from coalib.bears.GlobalBear import GlobalBear
 from coalib.results.RESULT_SEVERITY import RESULT_SEVERITY
 from coalib.results.Result import Result
-from coalib.bears.GlobalBear import GlobalBear
-from bears.codeclone_detection.ClangFunctionDifferenceBear import (
-    ClangFunctionDifferenceBear)
 
 
 class ClangCloneDetectionBear(GlobalBear):
