+++ bears/c_languages/codeclone_detection/ClangFunctionDifferenceBear.py
@@ -1,18 +1,18 @@
 import functools
 from itertools import combinations
 
+from bears.c_languages.codeclone_detection.ClangCountVectorCreator import (
+    ClangCountVectorCreator)
+from bears.c_languages.codeclone_detection.ClangCountingConditions import (
+    condition_dict)
+from bears.c_languages.codeclone_detection.CloneDetectionRoutines import (
+    compare_functions,
+    get_count_matrices)
+from coalib.bears.GlobalBear import GlobalBear
+from coalib.collecting.Collectors import collect_dirs
 from coalib.misc.StringConverter import StringConverter
 from coalib.results.HiddenResult import HiddenResult
 from coalib.settings.Setting import typed_ordered_dict, path_list
-from coalib.collecting.Collectors import collect_dirs
-from coalib.bears.GlobalBear import GlobalBear
-from bears.codeclone_detection.ClangCountVectorCreator import (
-    ClangCountVectorCreator)
-from bears.codeclone_detection.ClangCountingConditions import condition_dict
-from bears.codeclone_detection.CloneDetectionRoutines import (
-    compare_functions,
-    get_count_matrices)
-
 
 # counting_condition_dict is a function object generated by typed_dict. This
 # function takes a setting and creates a dictionary out of it while it
