+++ bears/tests/c_languages/codeclone_detection/CloneDetectionRoutinesTest.py
@@ -2,7 +2,7 @@ import sys
 import unittest
 
 sys.path.insert(0, ".")
-from bears.codeclone_detection.CloneDetectionRoutines import (
+from bears.c_languages.codeclone_detection.CloneDetectionRoutines import (
     relative_difference)
