+++ bears/tests/c_languages/codeclone_detection/CountVectorTest.py
@@ -1,9 +1,9 @@
-from math import sqrt
 import sys
+from math import sqrt
 
 sys.path.insert(0, ".")
 import unittest
-from bears.codeclone_detection.CountVector import CountVector
+from bears.c_languages.codeclone_detection.CountVector import CountVector
 
 
 class CountVectorTest(unittest.TestCase):
