+++ bears/tests/c_languages/codeclone_detection/clone_detection_samples/clones/one_big_one_small.c
