+++ bears/tests/c_languages/codeclone_detection/clone_detection_samples/clones/s2b.c
