+++ bears/tests/c_languages/codeclone_detection/clone_detection_samples/kernel_print.h
