+++ bears/tests/c_languages/codeclone_detection/clone_detection_samples/non_clones/builder_editorconfig.c
