+++ bears/tests/c_languages/codeclone_detection/conditions_samples.c
