+++ bears/tests/c_languages/codeclone_detection/sample.c
